interview_ai/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
jinja2==3.1.3
websockets==12.0
pydantic==2.6.1
python-dotenv==1.0.1
groq==0.4.2
opencv-python==4.8.0.74
mediapipe==0.10.8
numpy==1.24.3
pdfplumber==0.10.3
python-docx==1.1.0
sounddevice==0.4.6
scipy==1.11.3
librosa==0.10.1
aiofiles==23.2.1
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0
pydub==0.25.1 